Document upload working with guids for the names and db for the original names and respective information. Changed the type of the field SizeMB for the db. Composer to start the kafka usage.

+++ OpenBank.API/Domain/Models/Entities/Document.cs
@@ -16,7 +16,7 @@ public class Document : BaseEntity
     public string ContentType { get; set; }
 
     [Required]
-    public int SizeMB { get; set; }
+    public double SizeMB { get; set; }
 
     [Required]
     [ForeignKey("Accounts")]
